add tankoubon metadata write locks

+++ lib/LANraragi/Controller/Api/Tankoubon.pm
@@ -65,7 +65,25 @@ sub delete_tankoubon {
     my $self   = shift;
     my $tankid = $self->stash('id');
 
+    # lock resource
+    my $redis = LANraragi::Model::Config->get_redis;
+    my $lock = $redis->setnx( "tankoubon-write:$tankid", 1 );
+    if ( !$lock ) {
+        return $self->render(
+            json => {
+                operation => "delete_tankoubon",
+                success   => 0,
+                error     => "Locked resource: $tankid."
+            },
+            status => 423
+        );
+    }
+    $redis->expire( "tankoubon-write:$tankid", 10 );
+
     my $result = LANraragi::Model::Tankoubon::delete_tankoubon($tankid);
+
+    $redis->del( "tankoubon-write:$tankid" );
+    $redis->quit();
 
     if ($result) {
         render_api_response( $self, "delete_tankoubon" );
@@ -80,7 +98,25 @@ sub update_tankoubon {
     my $tankid = $self->stash('id');
     my $data   = $self->req->json;
 
+    # lock resource
+    my $redis = LANraragi::Model::Config->get_redis;
+    my $lock = $redis->setnx( "tankoubon-write:$tankid", 1 );
+    if ( !$lock ) {
+        return $self->render(
+            json => {
+                operation => "update_tankoubon",
+                success   => 0,
+                error     => "Locked resource: $tankid."
+            },
+            status => 423
+        );
+    }
+    $redis->expire( "tankoubon-write:$tankid", 10 );
+
     my ( $result, $err ) = LANraragi::Model::Tankoubon::update_tankoubon( $tankid, $data );
+
+    $redis->del( "tankoubon-write:$tankid" );
+    $redis->quit();
 
     if ($result) {
         my %tankoubon      = LANraragi::Model::Tankoubon::get_tankoubon($tankid);
@@ -98,7 +134,25 @@ sub add_to_tankoubon {
     my $tankid = $self->stash('id');
     my $arcid  = $self->stash('archive');
 
+    # lock resource
+    my $redis = LANraragi::Model::Config->get_redis;
+    my $lock = $redis->setnx( "tankoubon-write:$tankid", 1 );
+    if ( !$lock ) {
+        return $self->render(
+            json => {
+                operation => "add_to_tankoubon",
+                success   => 0,
+                error     => "Locked resource: $tankid."
+            },
+            status => 423
+        );
+    }
+    $redis->expire( "tankoubon-write:$tankid", 10 );
+
     my ( $result, $err ) = LANraragi::Model::Tankoubon::add_to_tankoubon( $tankid, $arcid );
+
+    $redis->del( "tankoubon-write:$tankid" );
+    $redis->quit();
 
     if ($result) {
         my $successMessage = "Added $arcid to tankoubon $tankid!";
@@ -121,7 +175,25 @@ sub remove_from_tankoubon {
     my $tankid = $self->stash('id');
     my $arcid  = $self->stash('archive');
 
+    # lock resource
+    my $redis = LANraragi::Model::Config->get_redis;
+    my $lock = $redis->setnx( "tankoubon-write:$tankid", 1 );
+    if ( !$lock ) {
+        return $self->render(
+            json => {
+                operation => "remove_from_tankoubon",
+                success   => 0,
+                error     => "Locked resource: $tankid."
+            },
+            status => 423
+        );
+    }
+    $redis->expire( "tankoubon-write:$tankid", 10 );
+
     my ( $result, $err ) = LANraragi::Model::Tankoubon::remove_from_tankoubon( $tankid, $arcid );
+
+    $redis->del( "tankoubon-write:$tankid" );
+    $redis->quit();
 
     if ($result) {
         my $successMessage = "Removed $arcid from tankoubon $tankid!";
